api/requirements.txt
fastapi==0.115.*
pyyaml==6.*
